refactored components and services
