+++ src/app/admin-login/admin-login.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Http, Response } from '@angular/http';
 import { Router } from '@angular/router';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { FormBuilder, FormGroup } from '@angular/forms';
 import { AdminLoginService } from '../services/admin-login-service/admin-login.service';
 
 @Component({
@@ -9,16 +8,14 @@ import { AdminLoginService } from '../services/admin-login-service/admin-login.s
   templateUrl: './admin-login.component.html',
   styleUrls: ['./admin-login.component.css']
 })
-export class AdminLoginComponent extends AdminLoginService implements OnInit {
+export class AdminLoginComponent implements OnInit {
 
-  private adminApiUrl = 'https://good-works-summer-backend.herokuapp.com/initialdata/admin';
   adminCredentials = [];
 
   adminForm: FormGroup;
   submitted = false;
 
-  constructor(private http: Http, private router: Router, private formBuilder: FormBuilder) {
-    super(http, formBuilder);
+  constructor(private router: Router, private formBuilder: FormBuilder, private adminLoginService: AdminLoginService) {
   }
 
   ngOnInit() {
@@ -28,7 +25,7 @@ export class AdminLoginComponent extends AdminLoginService implements OnInit {
   }
 
   getAdminCredentials() {
-    return this.getData(this.adminApiUrl);
+    return this.adminLoginService.getData(''); // apiURL laikinai tuscias
   }
 
 }
