+++ src/app/admin-login/admin-project/admin-project.component.ts
@@ -1,23 +1,24 @@
 import { Component, OnInit } from '@angular/core';
-import { AdminProjectService } from '../services/admin-project-service/admin-project.service';
-import { Http } from '@angular/http';
+import { AdminProjectService } from '../../services/admin-project-service/admin-project.service';
 import { Router } from '@angular/router';
 import { FormBuilder } from '@angular/forms';
+import { TEAMS_API_URL } from '../../registration.const';
+import { FILTERED_TEAMS_API_URL } from '../../registration.const';
+import { CATEGORY_API_URL } from '../../registration.const';
 
 @Component({
   selector: 'app-admin-project',
   templateUrl: './admin-project.component.html',
   styleUrls: ['./admin-project.component.css']
 })
-export class AdminProjectComponent extends AdminProjectService implements OnInit {
+export class AdminProjectComponent implements OnInit {
   readonly tableApiUrl = 'https://good-works-summer-backend.herokuapp.com';
   teams = [];
   categories = [];
   filteredTeams = [];
   private pathCategory: string;
 
-  constructor(private http: Http, private router: Router, private formBuilder: FormBuilder) {
-    super(http, formBuilder);
+  constructor(private router: Router, private formBuilder: FormBuilder, private adminProjectService: AdminProjectService) {
   }
 
   ngOnInit() {
@@ -30,15 +31,15 @@ export class AdminProjectComponent extends AdminProjectService implements OnInit
   }
 
   getTeams() {
-    return this.getData(this.tableApiUrl + '/teams');
+    return this.adminProjectService.getData(TEAMS_API_URL);
   }
 
   getCategories() {
-    return this.getData(this.tableApiUrl + '/initialdata/categories');
+    return this.adminProjectService.getData(CATEGORY_API_URL);
   }
 
   getFilteredTeams() {
-    return this.getData(this.tableApiUrl + '/teams/filter/' + this.pathCategory)
+    return this.adminProjectService.getData(FILTERED_TEAMS_API_URL + this.pathCategory)
   }
