+++ src/app/app.module.ts
@@ -14,7 +14,7 @@ import { MainPageComponent } from './main-page/main-page.component';
 import { HttpClientModule } from '@angular/common/http';
 import { JobRegistrationService } from './services/job-registration-service/job-registration.service';
 import { AdminLoginComponent } from './admin-login/admin-login.component';
-import { AdminProjectComponent } from './admin-project/admin-project.component';
+import { AdminProjectComponent } from './admin-login/admin-project/admin-project.component';
 
 const routes: Routes = [
   { path: '', component: MainPageCarouselComponent },
