+++ src/app/registration.const.ts
@@ -1,3 +1,5 @@
 export const ROOT_API_URL = 'https://good-works-summer-backend.herokuapp.com';
+export const TEAMS_API_URL = 'https://good-works-summer-backend.herokuapp.com/teams';
 export const CITIES_API_URL = 'https://good-works-summer-backend.herokuapp.com/initialdata/cities';
+export const FILTERED_TEAMS_API_URL = 'https://good-works-summer-backend.herokuapp.com/teams/filter';
 export const CATEGORY_API_URL = 'https://good-works-summer-backend.herokuapp.com/initialdata/categories';
